Fix price calculation
The computation script uses prices to iterate. If prices are placed in <div>s,
the browser renders them outside the table. When the row is then removed
upon paying, the div remains which causes an out-of-bounds error on the
next function call.
We use a <td> here to ensure that the element is rendere inside the
table.